scripts/requirements.txt
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
tqdm>=4.65.0